Refactor peer communication code into a separate file
Move peer communication related code to a new file `src/peer_communication.py`.

* **New File `src/peer_communication.py`**
  - Add `broadcast_block_header` function.
  - Add `receive_block_header` function.
  - Add `broadcast_transaction` function.
  - Import necessary modules and dependencies.

* **Changes in `src/tinychain.py`**
  - Remove `broadcast_block_header` function and its related code.
  - Remove `receive_block_header` function and its related code.
  - Remove `broadcast_transaction` function and its related code.
  - Import the new `peer_communication` module and use its functions.
